+++ core/products/perpetual.go
@@ -32,6 +32,8 @@ import (
 )
 
 var (
+	year = num.DecimalFromInt64((24 * 365 * time.Hour).Nanoseconds())
+
 	ErrDataPointAlreadyExistsAtTime = errors.New("data-point already exists at timestamp")
 	ErrInitialPeriodNotStarted      = errors.New("initial settlement period not started")
 )
@@ -304,8 +306,18 @@ func (p *Perpetual) handleSettlementCue(ctx context.Context, t int64) {
 	internalTWAP := twap(p.internal, p.startedAt, t)
 	// and calculate the same using the external oracle data-points over the same period
 	externalTWAP := twap(p.external, p.startedAt, t)
+
 	// do the calculation
 	fundingPayment, fundingRate := p.calculateFundingPayment(internalTWAP, externalTWAP)
+
+	// apply interest-rates if necessary
+	if !p.p.InterestRate.IsZero() {
+		delta := t - p.internal[0].t
+		if p.log.GetLevel() == logging.DebugLevel {
+			p.log.Debug("applying interest-rate with clamping", logging.String("funding-payment", fundingPayment.String()), logging.Int64("delta", delta))
+		}
+		fundingPayment.Add(p.calculateInterestTerm(externalTWAP, internalTWAP, delta))
+	}
 
 	// send it away!
 	fp := &num.Numeric{}
@@ -452,6 +464,36 @@ func (p *Perpetual) calculateFundingPayment(internalTWAP, externalTWAP *num.Uint
 		logging.String("funding-rate", fundingRate.String()))
 
 	return fundingPayment, &fundingRate
+}
+
+func (p *Perpetual) calculateInterestTerm(externalTWAP, internalTWAP *num.Uint, delta int64) *num.Int {
+	// get delta in terms of years
+	td := num.DecimalFromInt64(delta).Div(year)
+
+	// convert into num types we need
+	sTWAP := num.DecimalFromUint(externalTWAP)
+	fTWAP := num.DecimalFromUint(internalTWAP)
+
+	// interest = (1 + r * td) * s_swap - f_swap
+	interest := num.DecimalOne().Add(p.p.InterestRate.Mul(td)).Mul(sTWAP)
+	interest = interest.Sub(fTWAP)
+
+	upperBound := num.DecimalFromUint(externalTWAP).Mul(p.p.ClampUpperBound)
+	lowerBound := num.DecimalFromUint(externalTWAP).Mul(p.p.ClampLowerBound)
+
+	clampedInterest := num.MinD(upperBound, num.MaxD(lowerBound, interest))
+	if p.log.GetLevel() == logging.DebugLevel {
+		p.log.Debug("clamped interest and bounds",
+			logging.MarketID(p.id),
+			logging.String("lower-bound", p.p.ClampLowerBound.String()),
+			logging.String("interest", interest.String()),
+			logging.String("upper-bound", p.p.ClampUpperBound.String()),
+			logging.String("clamped-interest", clampedInterest.String()),
+		)
+	}
+
+	result, _ := num.IntFromDecimal(clampedInterest)
+	return result
 }
 
 // GetMarginIncrease returns the estimated extra margin required to account for the next funding payment.